requirements.txt
# FastAPI และ dependencies หลัก
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Google Sheets API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# Date and time handling
python-dateutil==2.8.2

# Logging และ monitoring
structlog==23.2.0

# Security
cryptography==41.0.8

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Performance monitoring (optional)
prometheus-client==0.19.0